+++ reverse-engineering/cimg-display-c/_11/cimg.c
@@ -66,7 +66,7 @@ struct cimg_header
     char magic_number[4];
     uint64_t version;
     uint16_t width;
-    uint64_t height;
+    uint16_t height;
 } __attribute__((packed));
 
 typedef struct
