+++ reverse-engineering/cimg-display-c/_4/cimg.c
@@ -64,7 +64,7 @@ void read_exact(int fd, void *dst, int size, char *msg, int exitcode)
 struct cimg_header
 {
     char magic_number[4];
-    uint16_t version;
+    uint64_t version;
     uint32_t width;
     uint32_t height;
 } __attribute__((packed));
