+++ reverse-engineering/cimg-display-c/_8/cimg.c
@@ -64,8 +64,8 @@ void read_exact(int fd, void *dst, int size, char *msg, int exitcode)
 struct cimg_header
 {
     char magic_number[4];
-    uint64_t version;
-    uint32_t width;
+    uint8_t version;
+    uint64_t width;
     uint16_t height;
 } __attribute__((packed));
